feat(vault): add animated border to the fab button
