+++ core/ui/src/commonMain/kotlin/com/thejohnsondev/ui/designsystem/Dimensions.kt
@@ -12,6 +12,8 @@ val Size8 = 8.dp
 val Size10 = 10.dp
 val Size12 = 12.dp
 val Size16 = 16.dp
+val Size19 = 19.dp
+val Size20 = 20.dp
 val Size24 = 24.dp
 val Size22 = 22.dp
 val Size32 = 32.dp
