+++ core/ui/src/commonMain/kotlin/com/thejohnsondev/ui/model/ScaffoldConfig.kt
@@ -16,5 +16,6 @@ data class ScaffoldConfig(
     val snackBarPaddingHorizontal: Dp? = null,
     val snackBarPaddingVertical: Dp? = null,
     val snackBarHostState: SnackbarHostState? = null,
-    val bottomBarItemIndex: Int = 0
+    val bottomBarItemIndex: Int = 0,
+    val isEmptyVaultScreen: Boolean = false
 )
